configs + requirements updated
