+++ pybullet_requirements.txt
@@ -6,3 +6,5 @@ stable-baselines3
 sb3-contrib @ git+https://github.com/RemiG3/stable-baselines3-contrib/@feat/ppo_attention
 torch
 pybullet
+PyYAML
+tensorflow
